+++ src/test/java/com/uplus/backend/integration/search/SearchIntegrationTest.java
@@ -1,36 +1,21 @@
-package com.uplus.backend.integration.device;
+package com.uplus.backend.integration.search;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.uplus.backend.device.entity.Color;
 import com.uplus.backend.device.entity.Device;
-import com.uplus.backend.device.entity.Image;
-import com.uplus.backend.device.entity.Tag;
-import com.uplus.backend.device.repository.ColorRepository;
 import com.uplus.backend.device.repository.DeviceRepository;
-import com.uplus.backend.device.repository.ImageRepository;
-import com.uplus.backend.device.repository.TagRepository;
-import com.uplus.backend.order.dto.OrderCreateRequestDto;
-import com.uplus.backend.order.repository.OrderRepository;
 import com.uplus.backend.plan.entity.Plan;
 import com.uplus.backend.plan.repository.PlanRepository;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.MethodOrderer;
-import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.TestInstance.Lifecycle;
-import org.junit.jupiter.api.TestMethodOrder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
@@ -38,24 +23,25 @@ import org.springframework.util.MultiValueMap;
 @SpringBootTest
 @AutoConfigureMockMvc
 @TestInstance(Lifecycle.PER_CLASS)
-public class PlanIntegrationTest {
+public class SearchIntegrationTest {
 
 	@Autowired
 	private MockMvc mockMvc;
+
+	@Autowired
+	private DeviceRepository deviceRepository;
 
 	@Autowired
 	private PlanRepository planRepository;
 
 	private Plan plan1;
 
-	private Plan plan2;
-
+	private Device device1;
 
 	@BeforeAll
 	private void setup() {
 		// given
 		plan1 = Plan.builder()
-			.id(1L)
 			.name("5G 시그니처")
 			.networkType(5)
 			.price(130_000)
@@ -63,33 +49,39 @@ public class PlanIntegrationTest {
 			.voiceCall("집/이동전화 무제한")
 			.message("기본제공")
 			.build();
-
 		plan1 = planRepository.save(plan1);
 
-		plan2 = Plan.builder()
-			.id(2L)
-			.name("5G 프리미어")
+		device1 = Device.builder()
+			.name("갤럭시  Z Fold 4 512GB")
+			.serialNumber("SM-F936N513")
+			.storage("512GB")
+			.price(2_119_700)
+			.launchedDate(LocalDateTime.now())
+			.company("삼성")
 			.networkType(5)
-			.price(90_000)
-			.data("60GB")
-			.voiceCall("집/이동전화 무제한")
-			.message("기본제공")
+			.cpu("스냅드래곤 8+Gen1 (SM8475) (4nm, Octa-Core)")
+			.display(
+				"메인 : 6.7” (170.3 mm) FHD+ , Dynamic AMOLED 2X 커버 : 1.9”(48.2 mm) , Super AMOLED")
+			.publicSupport(200_000)
+			.additionalSupport(50_000)
+			.repImageUrl(
+				"https://image.lguplus.com/static/pc-contents/images/prdv//20220812-025607-814-9q8rtAhk.jpg")
+			.plan(plan1)
 			.build();
-
-		plan2 = planRepository.save(plan2);
+		device1 = deviceRepository.save(device1);
 
 	}
 
-
 	@Test
-	void 요금제_조회_테스트() throws Exception {
+	void 검색_조회_테스트() throws Exception {
 		// given
 
 		MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
+		params.add("q", "갤럭시");
 		params.add("network-type", String.valueOf(5));
 
 		// when & then
-		mockMvc.perform(get("/api/plan").params(params))
+		mockMvc.perform(get("/api/search").params(params))
 			.andExpect(status().isOk())
 			.andDo(print());
 	}
